+++ lihil/plugins/interface.py
@@ -7,10 +7,6 @@ from lihil.signature import EndpointSignature
 
 
 class IEndpointInfo(Protocol, Generic[P, R]):
-    def __init__(
-        self, graph: Graph, func: IAsyncFunc[P, R], sig: EndpointSignature[R]
-    ): ...
-
     @property
     def graph(self) -> Graph: ...
     @property
@@ -20,8 +16,4 @@ class IEndpointInfo(Protocol, Generic[P, R]):
 
 
 class IPlugin(Protocol):
-    def __call__(
-        self,
-        endpint_info: IEndpointInfo[P, R],
-        /,
-    ) -> IAsyncFunc[P, R]: ...
+    def __call__(self, endpint_info: IEndpointInfo[P, R]) -> IAsyncFunc[P, R]: ...